Add simulation using Cholesky decomposition
The method used to compute the integrals in the first solution doesn't give a good approximation. Thus we cannot see accurately the trajectories smoothness when H is close to 1.
Cholesky decomposition is the more straightforward and direct method to simulate fBM sample paths but it has a large computational cost.